Refactor Dropio.Core.Drop.AddFile() to accept a FileUpload object
The FileUpload control available to ASP.NET developers does not provide
the developer with the file path (only the file name) so the current
implementation of AddFile() was useless to ASP.NET developers. AddFile()
will now take either a string with the file path or a FileUpload object.
Also, in ServiceAdapter.Instance.AddFile(), the FileStream object was
changed to a Stream object, as the FileUpload object does not provide a
method to return a FileStream object (but it will return a Stream
object). Changing the FileStream to a Stream was simpler than doing some
kludgy hack to get a FileStream from a FileUpload