Changes to methods for updating a drop
Changed Drop.Update() to use the the drop object itself to apply the
changes (change the properties of the Drop object then call Update()).
This will not work for the "Name" property (the API call is constructed
by grabbing data from the Drop object, so if the name has been changed a
non-existent (or worse another existing) drop will be named in the API
call.

The function Drop.ChangeName() was added to address this issue. This
current behavior is how the ruby client library currently operates
(which is the main motivating factor for these changes)