Changes to the way the library handles Asset and Drop objects when making certain api calls
-Properties whose values cannot be changed by a called to Update() or
Save() now have their getters declared as "internal" so that only the
client library is able to change them

-Asset.Update() now called Asset.Save() (same as ruby library)

-Asset.Save() now updates the Asset object by reloading it from the
asset response returned from an "asset update" api call. This helps to
ensure that the values in the object is correct.

-Drop.Destroy() and Asset.Destroy() now removes all values from the
objects. This helps prevent accidently accessing a drop/asset that no
longer exists (previously the values in the object were kept intact)

-Asset.CopyTo() now increases the asset count of the drop object that was
copied to

-Asset.MoveTo() now increases the asset count of the drop object that was moved
to, and decreases the asset count of the drop object that was moved from